requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.0
distro==1.4.0
Django==3.1.4
django-cities==0.6
django-countries==7.0
django-import-export==2.4.0
django-model-utils==4.1.1
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
jdcal==1.4.1
lockfile==0.12.2
MarkupPy==1.14
msgpack==0.6.2
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.3
pep517==0.8.2
Pillow==8.0.1
progress==1.5
pyparsing==2.4.6
python-decouple==3.3
python-dotenv==0.15.0
pytoml==0.1.21
pytz==2020.4
PyYAML==5.3.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
swapper==1.1.2.post1
tablib==3.0.0
tqdm==4.54.1
urllib3==1.25.8
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
